Change masterthesis to mastersthesis
The question which one is "correct" is apparently not completely trivial
to decide. Apparently Mendeley uses the former (at least this is implied
on <https://tex.stackexchange.com/q/415204>, but that reference also points
out this is wrong)

The wrong name is also given at <https://www.bibtex.com/e/entry-types/>
but note that the example they give here uses the right form, as do the
templates at <https://www.bibtex.com/t/template-mastersthesis/> (I've
reported this to the company running that website)

Also the biblatex manual lists only `mastersthesis`.

But ultimately, the source of truth should be bibtex itself; so see e.g.
<http://tug.org/svn/texlive/trunk/Build/extra/bibtex/alpha.bst?view=markup#l736>